Add files via upload
Changed Base EDA v1.0 INNO file to output the system.threading.dll file into the Landis Extension folder with the installer. This is the file necessary to do parallel processing that is missing from the .Net 3.5 Framework. Install file is the one created with the updated INNO so it installs the system.threading.dll file.